move class widget to shared folder
