+++ src/components/fields/schemaFields/CssClassField.tsx
@@ -17,7 +17,7 @@
 
 import React, { useMemo } from "react";
 import { defaultFieldFactory } from "@/components/fields/schemaFields/SchemaFieldContext";
-import CssClassWidget from "@/components/fields/schemaFields/widgets/CssClassWidget";
+import CssClassWidget from "@/components/fields/schemaFields/widgets/CssClassWidgets/CssClassWidget";
 import { getToggleOptions } from "@/components/fields/schemaFields/getToggleOptions";
 import { type SchemaFieldProps } from "@/components/fields/schemaFields/propTypes";
