+++ src/components/fields/schemaFields/widgets/CssClassWidgets/CssClassWidget.module.scss
@@ -1,4 +1,4 @@
-/*
+/*!
  * Copyright (C) 2023 PixieBrix, Inc.
  *
  * This program is free software: you can redistribute it and/or modify
@@ -14,13 +14,6 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-
-.spacingControlContainer {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 0.25rem;
-}
 
 .flagButton:global(.btn.active):not(:focus) {
   z-index: unset;
